requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
Django==3.0.5
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-template-maths==0.2.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
Pillow==7.1.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
stripe==2.48.0
urllib3==1.25.9
wrapt==1.11.2
